% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ise}
\alias{ise}
\title{Integrated squared error against a known density}
\usage{
ise(object, true_density)
}
\arguments{
\item{object}{A `"tiltdens"` or `"density"` object.}

\item{true_density}{A function of one numeric vector.}
}
\value{
A single number.
}
\description{
Approximates \eqn{\int (\hat f - f)^2} by the trapezoidal rule over the grid
of the fit. Useful for simulation studies where the truth is known.
}
\examples{
set.seed(1)
fit <- tilt_density(rnorm(80), m = 3)
ise(fit, dnorm)

}
