% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_registry.R
\name{plot.tilt_kernel}
\alias{plot.tilt_kernel}
\title{Plot a kernel}
\usage{
\method{plot}{tilt_kernel}(x, from = NULL, to = NULL, ...)
}
\arguments{
\item{x}{A `"tilt_kernel"` object.}

\item{from, to}{Range of the plot. Defaults to the support, or to four
standard deviations for an unbounded kernel.}

\item{...}{Passed to [graphics::plot()].}
}
\value{
`x`, invisibly.
}
\description{
Plot a kernel
}
\examples{
plot(tilt_kernel("biweight"))

}
