% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_fit.R
\name{plot.tiltdens_md}
\alias{plot.tiltdens_md}
\title{Contour plot of a bivariate fit}
\usage{
\method{plot}{tiltdens_md}(x, n = 60, points = TRUE, ...)
}
\arguments{
\item{x}{A `"tiltdens_md"` object with two dimensions.}

\item{n}{Grid points per axis.}

\item{points}{Draw the observations over the contours.}

\item{...}{Passed to [graphics::contour()].}
}
\value{
`x`, invisibly.
}
\description{
Contour plot of a bivariate fit
}
