% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/multivariate_fit.R
\name{predict.tiltdens}
\alias{predict.tiltdens}
\alias{predict.tiltdens_md}
\title{Evaluate a fitted density at new points}
\usage{
\method{predict}{tiltdens}(object, newdata = NULL, ...)

\method{predict}{tiltdens_md}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A `"tiltdens"` object.}

\item{newdata}{Numeric vector of points at which to evaluate. Defaults to the
grid stored in `object`.}

\item{...}{Ignored.}
}
\value{
A numeric vector of density values.
}
\description{
Evaluate a fitted density at new points
}
\examples{
set.seed(1)
fit <- tilt_density(rnorm(60), m = 3)
predict(fit, newdata = c(-1, 0, 1))

}
