% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.tiltdens}
\alias{print.tiltdens}
\title{Print a tilted or sharpened density fit}
\usage{
\method{print}{tiltdens}(x, digits = getOption("digits") - 2L, ...)
}
\arguments{
\item{x}{A `"tiltdens"` object.}

\item{digits}{Number of significant digits.}

\item{...}{Ignored.}
}
\value{
`x`, invisibly.
}
\description{
Print a tilted or sharpened density fit
}
