% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_fit.R
\name{print.tiltdens_md}
\alias{print.tiltdens_md}
\title{Print a multivariate fit}
\usage{
\method{print}{tiltdens_md}(x, digits = getOption("digits") - 2L, ...)
}
\arguments{
\item{x}{A `"tiltdens_md"` object.}

\item{digits}{Number of significant digits.}

\item{...}{Ignored.}
}
\value{
`x`, invisibly.
}
\description{
Print a multivariate fit
}
