% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltdens-package.R
\docType{package}
\name{tiltdens-package}
\alias{tiltdens}
\alias{tiltdens-package}
\title{tiltdens: Tilted and Data-Sharpened Nonparametric Density Estimation}
\description{
High-order nonparametric density estimators built by perturbing a conventional kernel estimator, either by re-weighting the observations ("tilting") or by moving them ("data sharpening"). The perturbation is chosen so that the estimator inherits the fast convergence rate of an infinite-order kernel estimator, such as the sinc or trapezoidal flat-top estimator, while remaining a proper non-negative density without the oscillatory tails those estimators suffer from. Two criteria are provided: minimising the L2 distance to an infinite-order comparator, following Doosti and Hall (2016) \doi{10.1111/rssb.12112}, and minimising a cross-validation criterion that needs no comparator and is much faster, following Doosti, Hall and Mateu (2018) \doi{10.1016/j.jspi.2017.12.003}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DoostiH/tiltdens}
  \item Report bugs at \url{https://github.com/DoostiH/tiltdens/issues}
}

}
\author{
\strong{Maintainer}: Hassan Doosti \email{hassan.doosti@mq.edu.au} (Author of the underlying methods)

Authors:
\itemize{
  \item Jorge Mateu
  \item Peter Hall (Author of the underlying methods; deceased 2016)
}

}
\keyword{internal}
